Adds GoodsNomenclatureHelper
The DeclarableHelper is meant for headings and commodities that are
declarable. Some of the functionality in there was really built for all
goods nomenclature so this extracts that behaviour into its own helper
