+++ app/controllers/trading_partners_controller.rb
@@ -1,5 +1,5 @@
 class TradingPartnersController < ApplicationController
-  include DeclarableHelper
+  include GoodsNomenclatureHelper
 
   before_action :disable_search_form, :disable_switch_service_banner do
     @tariff_last_updated = nil
