+++ app/views/meursing_lookup/steps/_end.html.erb
@@ -23,7 +23,7 @@
     <% if current_goods_nomenclature_code.present? %>
       <ul class="govuk-list govuk-list--bullet">
         <li>
-          <%= declarable_link %>
+          <%= goods_nomenclature_link %>
         </li>
         <li>
           <%= link_to('Start again', meursing_lookup_step_path(:start)) %>
